Move some stragglers.
