+++ Commands/FuncAssignCommand.cs
@@ -23,6 +23,7 @@
 using System;
 using System.Linq;
 using MetaphysicsIndustries.Solus.Expressions;
+using MetaphysicsIndustries.Solus.Functions;
 
 namespace MetaphysicsIndustries.Solus.Commands
 {
