+++ Macros/SqrtMacro.cs
@@ -26,6 +26,7 @@ using System.Linq;
 using System.Text;
 using MetaphysicsIndustries.Solus.Expressions;
 using MetaphysicsIndustries.Solus.Functions;
+using MetaphysicsIndustries.Solus.Macros;
 
 namespace MetaphysicsIndustries.Solus
 {
