+++ MetaphysicsIndustries.Solus.csproj
@@ -56,14 +56,15 @@
     <Compile Include="Expressions\DerivativeOfVariable.cs" />
     <Compile Include="Expressions\Expression.cs" />
     <Compile Include="Expressions\FunctionCall.cs" />
+    <Compile Include="Expressions\IExpressionVisitor.cs" />
     <Compile Include="Expressions\Literal.cs" />
+    <Compile Include="Expressions\ProductExpression.cs" />
     <Compile Include="Expressions\RandomExpression.cs" />
     <Compile Include="Expressions\SolusMatrix.cs" />
     <Compile Include="Expressions\SolusTensor.cs" />
     <Compile Include="Expressions\SolusVector.cs" />
     <Compile Include="Expressions\StringExpression.cs" />
     <Compile Include="Expressions\VariableAccess.cs" />
-    <Compile Include="Function.cs" />
     <Compile Include="Functions\AbsoluteValueFunction.cs" />
     <Compile Include="Functions\AdditionOperation.cs" />
     <Compile Include="Functions\ArccosecantFunction.cs" />
@@ -91,6 +92,7 @@
     <Compile Include="Functions\ExponentOperation.cs" />
     <Compile Include="Functions\FactorialFunction.cs" />
     <Compile Include="Functions\FloorFunction.cs" />
+    <Compile Include="Functions\Function.cs" />
     <Compile Include="Functions\GreaterThanComparisonOperation.cs" />
     <Compile Include="Functions\GreaterThanOrEqualComparisonOperation.cs" />
     <Compile Include="Functions\IfFunction.cs" />
@@ -113,14 +115,15 @@
     <Compile Include="Functions\TangentFunction.cs" />
     <Compile Include="Functions\UnaryOperation.cs" />
     <Compile Include="Functions\UnitStepFunction.cs" />
+    <Compile Include="Functions\UserDefinedFunction.cs" />
     <Compile Include="IPerPixelOperator.cs">
       <SubType>Code</SubType>
     </Compile>
-    <Compile Include="Macro.cs" />
     <Compile Include="Macros\AssignMacro.cs" />
     <Compile Include="Macros\DeleteMacro.cs" />
     <Compile Include="Macros\DeriveMacro.cs" />
     <Compile Include="Macros\FeedbackMacro.cs" />
+    <Compile Include="Macros\Macro.cs" />
     <Compile Include="Macros\RandMacro.cs" />
     <Compile Include="Macros\SqrtMacro.cs" />
     <Compile Include="Macros\SubstMacro.cs" />
@@ -139,7 +142,6 @@
     <Compile Include="Transformers\TransformArgs.cs" />
     <Compile Include="SolusAction.cs" />
     <Compile Include="OperationPrecedence.cs" />
-    <Compile Include="ProductExpression.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="SolusEngine.cs" />
     <Compile Include="SolusEngine.GatherVariables.cs" />
@@ -149,13 +151,11 @@
     <Compile Include="SolusParser.cs" />
     <Compile Include="SolusGrammar.cs" />
     <Compile Include="SolusEnvironment.cs" />
-    <Compile Include="IExpressionVisitor.cs" />
     <Compile Include="SingleHelper.cs" />
     <Compile Include="Instruction.cs" />
     <Compile Include="Transformers\VariableTransformArgs.cs" />
     <Compile Include="VariableToArgumentNumberMapper.cs" />
     <Compile Include="STuple.cs" />
-    <Compile Include="UserDefinedFunction.cs" />
     <Compile Include="MemoryImage.cs" />
   </ItemGroup>
   <ItemGroup>
